+++ DelvUI/Config/Tree/ConfigPageNode.cs
@@ -2,6 +2,7 @@ using Dalamud.Logging;
 using DelvUI.Config.Attributes;
 using DelvUI.Config.Profiles;
 using DelvUI.Helpers;
+using DelvUI.Interface.GeneralElements;
 using ImGuiNET;
 using Newtonsoft.Json;
 using System;
@@ -28,6 +29,50 @@ namespace DelvUI.Config.Tree
                 GenerateNestedConfigPageNodes();
                 _drawList = null;
             }
+        }
+
+        public override List<T> GetObjects<T>()
+        {
+            List<T> list = new List<T>();
+            if (_configObject == null) { return list; }
+
+            Type type = typeof(T);
+            if (_configObject is T obj)
+            {
+                list.Add(obj);
+            }
+
+            PropertyInfo[] properties = _configObject.GetType().GetProperties();
+            foreach (PropertyInfo property in properties)
+            {
+                if (property.PropertyType != type && !property.PropertyType.IsSubclassOf(type))
+                {
+                    continue;
+                }
+
+                object? value = property.GetValue(_configObject);
+                if (value != null && value is T o)
+                {
+                    list.Add(o);
+                }
+            }
+
+            FieldInfo[] fields = _configObject.GetType().GetFields();
+            foreach (FieldInfo field in fields)
+            {
+                if (field.FieldType != type && !field.FieldType.IsSubclassOf(type))
+                {
+                    continue;
+                }
+
+                object? value = field.GetValue(_configObject);
+                if (value != null && value is T o)
+                {
+                    list.Add(o);
+                }
+            }
+
+            return list;
         }
 
         private void GenerateNestedConfigPageNodes()